requirements.txt
ffmpeg==1.4
ffmpeg_python==0.2.0
keras==2.12.0
librosa==0.9.2
matplotlib==3.6.2
numpy==1.23.4
openai==0.27.6
opencv_python==4.7.0.72
Pillow==9.5.0
Pillow==9.5.0
PyAudio==0.2.13
setuptools==49.2.1
SpeechRecognition==3.10.0
tensorflow==2.12.0
tensorflow_intel==2.12.0
tqdm==4.64.1
transformers==4.28.1
tk